The complete node is complete - initially.
What a pain. This node has soooo many edge cases.

It seems so innocent and simple but its not. It poses the question: when
is a node complete? A join node is never complete, a switch node is never
complete, a split node ... etc

Even a complete node can be complete.

This initial implementation comes with plenty of test cases that hopefully
come some of the edge cases.