Add a release status page
For now, it appears in the "server root" as a means to help
direct people where the application lives.

With the addition of templates and an over-engineered structure,
it should be possible to create additional non-application static
pages with relative ease.

If/Once a good alternative "index" page is developed, the inane
technical jabber can move to "status" or whatever.