Refactored the receivership module.
Using the gen_server pattern for this does not appear to work since
node types as gen_server means that when on node is stopped, all nodes
of the same type are also stopped. At least that's what it appeared like
to me when I was testing using gen_server instead of receivership.

There has to be a better way.