Major refactoring of the way nodes and the system interact.
This change introduces the concept of "events" (!!) that nodes can
now receive and handle - handle using the `handle_event/2` function.

Events represent messages from the system to the node(s). Something like a
deploy is a system "event" or registration after initialisation. Or stop when
the system stops an entire flow.

In addition nodes receive "messages" from other nodes, these are then
incoming, outgoing and node specific messages.

Of course events and messages are both just Erlang messages, but its nice
to differentiate conceptually between types of messages that nodes may
or may not receive.