Initial final refactoring of the nodes and how these are structured.
This commit introduces the ered_node behaviour and throws out the receivership
module that was handling message routing between nodes and between nodes and
the system.

The ered_node behaviour is based on the gen_server behaviour and provides
a boilderplate for all nodes to use so that node implementations can focus
solely on functionality.

Each node is expected to implement three functions

- start for initialisation of the nodes behaviour
- handle_event for handling system messages
- handle_msg for handling node messages

Both handle functions are async, with the system messages being handle_info
messages and the node communication message being handle_cast messages.

Nodes do not have supervisors as yet but since they are now based on
gen_server, adding supervisors (for example one per flow) shouldn't be too
hard.

Much effort has been made to think this through but nothing is perfect, so
this might change but only to be extended with extra features.