Example Flow: Http Routing
First serious example using Erlang-RED: using the http in node to route
traffic for various http requests.

Once the flow has been deployed, it cam be accessed via `http://host-of-erlang-red/example01/page1`

It provides a very simple routing example but that can be expanded out to create
a blog using Erlang-RED.

The flow also has testing nodes that are triggered by an inject node which
is triggered by a unit test in the flow editor.

This allows the flow to demonstrate that routing http traffic is possible and
at the same time the same flow can test itself to ensure that it works.