A very stupid JSONata parser for Erlang!
But crazy ideas start small only to either fail or become big, either way
it's worth a try.

The intention of the JSONata parser is to transform JSONata into Erlang code
that can then be executed either at runtime or precompiled before executing
the flow code.

Also the set of valid JSONata stanzas that will be convertible will be the set
that I need[1] - this being a very much smaller set than all valid JSONata[2].

I'm sure this approach is a deadend since JSONata can become extremely
complex and conversion to Erlang will be messy and complex but its an approach
that can be quickly and dirtily implemented.

Plus, in the process, the learnings of playing around with leex and yecc are
well worth it: niche concepts (parser construction), in a niche programming
language (Erlang) for a niche problem space (JSONata) and the whole thing
applied to a niche development concept (visual FBP)! :)

[1]: https://github.com/gorenje/erlang-red/blob/40b489e9491d6f1ef38218838e21c0d85f6f0e30/priv/jsonata/leex.examples.json
[2]: https://github.com/jsonata-js/jsonata/pull/731#issuecomment-2593730718