JSONata executor for evaluating JSONata stanzas in the context of a Msg object.
This provides the scaffolding for using JSONata statements in the change,
inject and switch nodes although this hasn't yet been added.

The JSONata parser is a little bit better and can handle more statements
but it remains a very basic implementatation that can handle 1) map creation,
2) string concatenation and 3) basic JSONata function execution (replace and count ;))

Much to improve but the basics are in place.