Link Nodes: support dynamic link calls to link in nodes by either name or node id. Initial, semi-tested version of dynamic link calls.
Why is this important?

Because it allows for dynamic routing of http requests based on path parameters
for example. Each path in a blog represents different content but instead of
statically link paths and content, this can be done dynamically by having
a link-in naming convention, i.e., something like "[blog] <path>" where the
path is the URL path.

If the link-in node with a specific name does not exist, then it's a 404 error.