MQTT node: initial implementation.
mqtt in and mqtt out nodes provide an initial interface to an MQTT broker.
This is now allows a Node-RED instance to communicate with an Erlang-RED
instance :)

Further testing and development is needed but the general architecture is done.
This includes a gen_server wrapper around emqtt client. That prevents the
emqtt client from bring down a bunch of processes if it fails. The wrapper
also maps the generic event message names from emqtt to distinct MQTT
messages within the Erlang-RED internals.

Also config nodes have their place amongst all the other nodes. This allows
for configuration of a MQTT connection via the flow editor.