Exec node: initial implementation.
This is done much like the MQTT nodes with a "manager" module inbetween the
node and the library (erlexec) actually doing the work.

At this stage, the exec node supports spawn mode and killing processes.
Appending strings to command lines and adding values from the msg object
is also not supported.