heroku deployment: using herokus docker container stack.
The deployment is an initial effort to have a single flow executing without
a flow editor to make modifications.

The flow[1] in this case has a http-in node that accepts a request on '/'
and returns a single page --> https://red-erik.org is the result.

ErlangRED is executing that flow or rather ErlangRED created the processes
the respond to the web request.

[1] = https://github.com/gorenje/erlang-red/blob/main/priv/testflows/flow.499288ab4007ac6a.json