Added FlowHub.org
FlowHub is a platform for sharing and managing visual code *visually*.

Adding the FlowHub nodes to the project so that I can easily maintain the
git repo of flow testsuite --> https://github.com/gorenje/erlang-red-flow-testsuite

Also FlowHub runs on NodeRED and ErlangRED so it will be ideal for sharing code
between the two. In fact this implementation does not require a backend, it
all works in the frontend so that it can be used for other projects that want
to port NodeRED to other programming languages.

Full disclosure: I happened to have developed FlowHub out of a need to have
a proper visual code comparison, version management and sharing service. Also
FlowHub is completing implemented in NodeRED so its dogfood all the way down.