CSV Node: initial first completion of the decoder part of the node.
This introduces the ered_csv_parser_store gen(eric) server for storing and
sharing various parser types - differentiated by their separators.

The tests cover a fair good range of scenarios, but as I learnt in creating
them, there is always something else that is broken. Switch node, assert values
node and JSONata parser were all touched to get this done.

The csv parser store also allows for swapping out the underlying library for
CSV support, at the moment, it's nimble_csv which is an Elixir library.

Once all the tests are completed, will try on hnc-csv which is an Erlang
library - nice to have the choice!