Supervisor node: remove exception handling from the node behaviour.
With the introduction of a supervisor node, the handling of node failures
is left to the user. The idea of ErlangRED handling node failures does not
fit to Erlang.

It is now up to individual nodes to handle exceptions and post those back
to the system, alternatively the user can configure a supervisor that captures
process failure and restart or not the corresponding process.

Since NodeJS does not have the concept of process failure as a manageable
failure, there isn't a supervisor node in Node-RED. Instead Node-RED defines
the catch node that captures exception raised by the nodes. So each node knows
when it has to bubble up an exception to an exception handler. An exception in
Node-RED itself is not handled by a catch node, Node-RED just dies.

Erlang has processes that can either have exceptions or die. Supervisors are
processes that restart processes that die. Exceptions are catchable by the
process, process death is final not manageable by the process itself.

Therefore ErlangRED needs to add a Supervisor node that defines the
supervision of nodes contained in a flow. This can only happen if nodes are
allowed to die. Hence this change: ErlangRED allows processes and nodes to
die and be resurrected by users if this is desired.