supervisor node: send status messages on state change
Allow supervisor nodes to be restarted by other flow components by checking
the status messages sent by the supervisor node.

This is ironically a mixing of Node-RED concepts and Erlang concepts. The
supervisor node is responsible for restarting processes but itself is restarted
by a message sent to it from another node.

Essentially what should happen is that a further supervisor ensures that when
a supervisor crashes, it is restarted. Supervisors are responsible for
restarting nodes.

This solution goes against that. But that's because I haven't implemented the
supervisor-of-supervisor functionality yet. The way supervisors are implemented
at the moment, makes supervisor-of-supervisor impossible.