Live message tracing of flows.
Sometimes it is hard to visual the flow of message between nodes. Especially in
live systems when something goes wrong. The normal workflow would be to add
debug nodes and redeploy the flows and hope that the bug/errors/event happens
again.

This feature enables dynamic *without* redeploying the flows debugging of message
flows. Message tracing can be activated on a live system.

Once activated, messages passed to nodes (incoming messages are traced) can be
sent to the debug panel and/or visualised within the flow editor. Each node
that receives a message has a status change "msg received". Nodes are also
listed in panel in order of message receivership.

Additionally the contents of message can be sent to the debug panel for further
debugging.

More details at the package description --> https://flows.nodered.org/node/@gregoriusrippenstein/node-red-contrib-introspection

This feature was originally developed for Node-RED but has been ported to
Erlang-Red to improve the debugging workflow for flows.