Function node timeout implementation.
This implements the timeout feature of a function node and it has many
edgecases. What happens if timeout expires and the node is being supervised
by a supervisor mode? What happens with subprocesses when a function node
gets killed by a supervisor node? What happens when the func code is broken?

Basically the function ndoe is a bit jekyll & hyde: it behaviours slightly
differently when supervised than when it's not being supervised. The reason
for this is that processes for nodes should not go down unless they are
managed by supervisor nodes. But a function node creates two subprocesses per
message and normally if these subprocesses fail, the node process does not
care - it remains up. Processes for nodes don't die.

But that's not the case when the node is being supervised. That is because
the supervisor node supervises the process for the node, not the subprocesses
that the function node might or might not create. So to indicate to the
supervisor node that something went wrong, the node process has to die.

It gets even worse if the supervisor has a restart policy of "one for all"
or "rest for one" - the supervisors killing innocent function nodes just
because a sibling node was malfunctioning. So how to stop the healthy
subprocesses of a healthy function node if an unhealthy third party node fails
and the supervisor needs to cleanup?

Solution is quite simple: each subprocess of a function node monitors the
node process. Node process goes down, the subprocess gets a message and also
dies. Of course it goes in both direction: subprocess unexpectedly dies, the
node process also gets informed and decides what to do.

All this is tested by a bunch of "unit flow tests" for the function and
supervisor nodes:

- https://flows.red-erik.org/f/f2f61ec9fc46a468
- https://flows.red-erik.org/f/49c0a3ed69d61f56
- https://flows.red-erik.org/f/442b3bf0d630e21d
- https://flows.red-erik.org/f/d97b4eeb4ce93f0b
- https://flows.red-erik.org/f/c5f513bc380bc30a