Sort loaded applications for status
My friend said it was hard for him to find the packages in which he was
interested (mostly `erlang_red*`.  I was so excited that anyone was even
looking at the site that I made this change.

In the long term, if this page stays as-is and the overhead of doing all
of this dynamically becomes an issue, I'll work to build the structures
at compile time.