Support Node-RED property access including array indicies.
Node-RED is very liberal on how nested objects and arrays are accessed or
defined, allowing for a mix of array indicies and object key names.

For example

    key.12.key.12.12 = "fubar"

will generate the following structure:

```
   { "key": [
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        null,
        {
            "key": [
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                null,
                {
                    "12": "fubar"
                }
            ]
        }
      ]
    }
```

which is incredible useful for the end user, just extremely painful to implement
for the developer.

Because the propery can also be specified as `key[12]key.12["12"]` or a hundred
different combinations of array access, dot access and single and double quotes.

Also Node-RED has the convention that the last index (as above) will be
interpreted not as an array but as a keyname, i.e 12 becomes "12" and the key
for the object containing fubar. Why this is so, I don't know.

Anyway, in my endeavours to emulate Node-RED, this behaviour has to be implemented
aswell.

Either way, Erlang-Red can do all these fine tricks now too plus (and here comes
the artists freedom to innovate), any key in single quotes is converted to an
Erlang atom. Else strings are converted to binaries but if external data comes
in with atom keys, these can be addressed using single quotes.