No more floweditor tips.
Erlang-Red is not really for folks who have not worked with Node-RED before.

For an introduction into visual programming using Node-RED, checkout
blog.openmindmap.org (for example) or head over to the forum - discourse.nodered.org

If you are unconvinced and want tips within the flow editor, go to the settings
and turn them on there.